% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_cores}
\alias{max_cores}
\title{Maximum Number of OpenMP Threads}
\usage{
max_cores()
}
\value{
An \code{integer} with the maximum number of \emph{OpenMP} threads, or \code{0} when the
package was built without \emph{OpenMP} support.
}
\description{
Reports the number of threads \emph{OpenMP} would use on this machine.
}
\examples{
max_cores()
}
