% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_numeric_unif}
\alias{draw_numeric_unif}
\title{Static Thread Local Parallelized Uniform RNG sampler}
\usage{
draw_numeric_unif(n, min = 0, max = 1)
}
\arguments{
\item{n}{Number of values to generate.}

\item{min}{Bottom value to generate values from in a continuous uniform.}

\item{max}{Top range to generate values from in a continuous uniform.}
}
\value{
A \code{numeric} vector containing the output.
}
\description{
Samples from the Uniform distribution in parallel through using
core specific RNG engines that are kept active throughout the calculation.
}
\examples{
# Retrieves 100 observations from a random uniform distribution.
x = draw_numeric_unif(100)
}
