% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_cores}
\alias{max_cores}
\title{Max CPU cores}
\usage{
max_cores()
}
\value{
Integer containing the maximum number of CPU cores.

If 0, then parallelization is not supported and a silent failure occurs
}
\description{
Determine the number of cores available
}
\examples{

# Retrieve the number of total cores on the machine.
max_cores()
}
