% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppRNGParallel-package.R
\docType{package}
\name{RcppRNGParallel-package}
\alias{RcppRNGParallel}
\alias{RcppRNGParallel-package}
\title{RcppRNGParallel: Random Number Generation in Parallel with Static Threads}
\description{
Explores the best practices for parallel number generation streams.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
